+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Filter, search and sort the event log #376 #455
 - Dark mode! #420 #453
 - Support multiple wallets: WalletConnect, Trust, Rainbow etc #424
 - Export the user list as csv #402 #450